+++ Usage.md
@@ -1,0 +1,10 @@
+
+
+# Algorithem #
+
+To test the implementation of the algorithem, run test domain, which simulates the drone understanding of the space decompositon, this effecivly emulates the project without a physical drone
+
+
+# Drone #
+
+To test the implementation of the Drone, run DroneMainFile.py. This will require the crazyflie 2.0 or 2.1 with a flow deck and multiranger deck.
